+++ include/core/parameters_multiphysics.h
@@ -14,8 +14,6 @@
 #include <deal.II/base/parameter_handler.h>
 #include <deal.II/base/utilities.h>
 
-#include <algorithm>
-
 using namespace dealii;
 
 namespace Parameters
@@ -141,6 +139,30 @@ namespace Parameters
     electric_field,
     magnetic_field,
     power
+  };
+
+  /**
+   * @brief Parameters for the time coupling strategy of the time-harmonic Maxwell solver with the other physics solvers.
+   * - none: the time-harmonic Maxwell solver will only be solved at the first
+   * iteration of the time integration and will not be solved in the subsequent
+   * iterations.
+   * - iteration: the time-harmonic Maxwell solver will be solved every N time
+   * iterations, where N is given by the user input parameter "coupling
+   * iteration".
+   * - time: the time-harmonic Maxwell solver will be solved when crossing a
+   * time multiple, e.g. every T seconds, where the time multiple is given by
+   * the user input parameter "coupling time".
+   * - threshold: the time-harmonic Maxwell solver will be solved when the
+   * maximum change in the material properties of the problem change beyond a
+   * certain threshold, where the threshold value is given by the user input
+   * parameter "coupling threshold". **This is not implemented yet.**
+   */
+  enum class TimeHarmonicMaxwellCouplingStrategy : std::int8_t
+  {
+    none,
+    iteration,
+    time,
+    threshold
   };
 
   /**
@@ -433,6 +455,27 @@ namespace Parameters
   template <int dim>
   struct TimeHarmonicMaxwell
   {
+    // Integration method for the time coupling of the time-harmonic Maxwell
+    // solver
+    Parameters::TimeHarmonicMaxwellCouplingStrategy time_coupling_strategy;
+
+    // Coupling parameter for the time coupling strategy. If the time coupling
+    // strategy is based on the iteration count, this parameter represents the
+    // number of time iterations between two consecutive resolutions of the
+    // electromagnetic fields.
+    unsigned int coupling_iteration;
+
+    // If the time coupling strategy is based on time, this parameter represents
+    // the real time interval between two consecutive resolutions of the
+    // electromagnetic fields.
+    double coupling_time;
+
+    // If the time coupling strategy is based on a threshold, this parameter
+    // represents the change in the electromagnetic properties of the medium
+    // (i.e., permittivity, permeability or conductivity) that triggers the
+    // recomputation of the electromagnetic fields.
+    double coupling_threshold;
+
     // We use vectors in the following to be able to define multiple waveguides
     // in the same simulation.
     unsigned int number_of_waveguide_inlets;
@@ -444,8 +487,8 @@ namespace Parameters
     // to the right boundary
     std::vector<int> waveguide_boundary_ids;
 
-    // Waveguide power (in W) to be able to link the amplitude of the
-    // electromagnetic wave to the power injected in the waveguide
+    // Waveguide power (in W, time-averaged) to be able to link the amplitude
+    // of the electromagnetic wave to the power injected in the waveguide
     std::vector<double> waveguide_power;
 
     // Electric field amplitude used for the normalization of the solution
@@ -482,7 +525,6 @@ namespace Parameters
     parse_parameters(ParameterHandler &prm, const Dimensionality &dimensions);
   };
 
-
   /**
    * @brief Multiphysics - the parameters for multiphysics simulations
    * and handles sub-physics parameters.
@@ -500,6 +542,7 @@ namespace Parameters
     // subparameters for heat_transfer
     bool viscous_dissipation;
     bool thermal_buoyancy_force;
+    bool microwave_heating;
 
     Parameters::CLS                      cls_parameters;
     Parameters::CahnHilliard             cahn_hilliard_parameters;