+++ include/solvers/auxiliary_physics.h
@@ -203,6 +203,15 @@ public:
    */
   void
   setup_preconditioner() override{};
+
+  /**
+   * @brief Manage if this auxiliary physics should be solved at the current iteration. By default, auxiliary physics are solved at every iteration.
+   */
+  virtual bool
+  should_solve_auxiliary_physics()
+  {
+    return true;
+  }
 };